AgenticAI/requirements.txt
langgraph==0.2.39
langchain>=0.3.27,<2.0.0
langchain-core>=0.3.78,<2.0.0
langchain-openai>=0.2.6
langchain-community>=0.3.31
openai==1.54.0
pydantic==2.9.2
python-dotenv==1.0.1
tavily-python==0.3.5
requests>=2.32.5,<3.0.0
numpy>=2.1.0
rich==13.9.4
httpx>=0.27.0,<0.28.0
